requirements.txt
psycopg2
mysql-connector-repackaged
